+++ test/unit/cli/options_test.dart
@@ -4,45 +4,50 @@ import 'package:dart_test_tools/test.dart';
 import 'package:logging/logging.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:podman_backup/src/adapters/environment_adapter.dart';
+import 'package:podman_backup/src/adapters/posix_adapter.dart';
 import 'package:podman_backup/src/cli/options.dart';
-import 'package:podman_backup/src/models/hook.dart';
 import 'package:test/test.dart';
-import 'package:tuple/tuple.dart';
 
 class MockEnvironmentAdapter extends Mock implements EnvironmentAdapter {}
 
+class MockPosixAdapter extends Mock implements PosixAdapter {}
+
 void main() {
   group('$BackupMode', () {
-    testData<Tuple2<BackupMode, bool>>(
+    testData<(BackupMode, bool)>(
       'backup is mapped correctly',
       const [
-        Tuple2(BackupMode.full, true),
-        Tuple2(BackupMode.backupOnly, true),
-        Tuple2(BackupMode.uploadOnly, false),
+        (BackupMode.full, true),
+        (BackupMode.backupOnly, true),
+        (BackupMode.uploadOnly, false),
       ],
       (fixture) {
-        expect(fixture.item1.backup, fixture.item2);
+        expect(fixture.$1.backup, fixture.$2);
       },
     );
 
-    testData<Tuple2<BackupMode, bool>>(
+    testData<(BackupMode, bool)>(
       'upload is mapped correctly',
       const [
-        Tuple2(BackupMode.full, true),
-        Tuple2(BackupMode.backupOnly, false),
-        Tuple2(BackupMode.uploadOnly, true),
+        (BackupMode.full, true),
+        (BackupMode.backupOnly, false),
+        (BackupMode.uploadOnly, true),
       ],
       (fixture) {
-        expect(fixture.item1.upload, fixture.item2);
+        expect(fixture.$1.upload, fixture.$2);
       },
     );
   });
 
   group('$Options', () {
     final mockEnvironmentAdapter = MockEnvironmentAdapter();
+    final mockPosixAdapter = MockPosixAdapter();
 
     setUp(() {
       reset(mockEnvironmentAdapter);
+      reset(mockPosixAdapter);
+
+      when(() => mockPosixAdapter.isRoot).thenReturn(false);
     });
 
     group('sets correct backupCache defaults', () {
@@ -50,7 +55,10 @@ void main() {
         when(() => mockEnvironmentAdapter['HOME'])
             .thenReturn('/home/test-user');
 
-        final parser = Options.buildArgParser(mockEnvironmentAdapter);
+        final parser = Options.buildArgParser(
+          mockEnvironmentAdapter,
+          mockPosixAdapter,
+        );
 
         final backupCacheOption = parser.options['backup-cache'];
         expect(backupCacheOption, isNotNull);
@@ -63,7 +71,10 @@ void main() {
       test('sets correct path without HOME', () {
         when(() => mockEnvironmentAdapter['HOME']).thenReturn(null);
 
-        final parser = Options.buildArgParser(mockEnvironmentAdapter);
+        final parser = Options.buildArgParser(
+          mockEnvironmentAdapter,
+          mockPosixAdapter,
+        );
 
         final backupCacheOption = parser.options['backup-cache'];
         expect(backupCacheOption, isNotNull);
@@ -74,58 +85,23 @@ void main() {
       });
     });
 
-    testData<(List<String>, Map<String, Hook>)>(
-      'Can parse volume hooks correctly',
+    testData<(bool, bool)>(
+      'sets correct user defaults',
       const [
-        (
-          [],
-          {},
-        ),
-        (
-          ['volume-1=service1.service'],
-          {
-            'volume-1': Hook(unit: 'service1', type: 'service'),
-          },
-        ),
-        (
-          [
-            'volume-2=!service2.service',
-            'volume-3=service3@.service',
-          ],
-          {
-            'volume-2': Hook(
-              unit: 'service2',
-              type: 'service',
-              preHook: true,
-            ),
-            'volume-3': Hook(
-              unit: 'service3',
-              type: 'service',
-              isTemplate: true,
-            ),
-          },
-        ),
-        (
-          ['volume-4=!service4@.container'],
-          {
-            'volume-4': Hook(
-              unit: 'service4',
-              type: 'container',
-              isTemplate: true,
-              preHook: true,
-            ),
-          },
-        ),
+        (false, true),
+        (true, false),
       ],
       (fixture) {
-        final args = [
-          for (final arg in fixture.$1) '-H$arg',
-        ];
+        when(() => mockPosixAdapter.isRoot).thenReturn(fixture.$1);
 
-        final parser = Options.buildArgParser(mockEnvironmentAdapter);
-        final options = Options.parseOptions(parser.parse(args));
+        final parser = Options.buildArgParser(
+          mockEnvironmentAdapter,
+          mockPosixAdapter,
+        );
 
-        expect(options.getVolumeHooks(), fixture.$2);
+        final userOption = parser.options['user'];
+        expect(userOption, isNotNull);
+        expect(userOption!.defaultsTo, fixture.$2);
       },
     );
 
@@ -137,7 +113,10 @@ void main() {
           if (fixture != null) '-L${fixture.name.toLowerCase()}',
         ];
 
-        final parser = Options.buildArgParser(mockEnvironmentAdapter);
+        final parser = Options.buildArgParser(
+          mockEnvironmentAdapter,
+          mockPosixAdapter,
+        );
         final options = Options.parseOptions(parser.parse(args));
 
         expect(options.logLevel, fixture ?? Level.INFO);